Switch to multiversioned kernel packages
DNF supports multiple versions of the same package name
being installed as long as there are no file conflicts, and
handles responsibly cleaning up old kernels automatically
when this functionality is used properly.

Since we are exclusively using DNF for package management,
we should use this functionality so that kernels are more
manageable on systems.