Drop modXYZ for good
The kernel has this options itself to compress module, there is no need to do that manually.
Added an patch to compress the mods with ZSTD from the kernel since our kmod supports ZSTD.
If an an ARCH doesn't support ZSTD. set XZ or GPIP for the compression in the config
and add a BR on whatever is needed.